Send maximum possible frame data if it's going to overflow
There is a bug in RTP that limits the frame size to UINT16_MAX bytes, but a
frame migh in fact exceed this size limit, and when coverting the actual size
to uint16_t an overflow occurs, resulting in us sending a small portion of the
frame or nothing at all, if the size overflows to be 0. A proper fix would
likely break the toxav protocol, but we can improve the current behaviour
without breaking the protocol. We detect when the overflow occurs and instead of
letting it happen we send as much of the frame data as we can, avoiding the
cases when we would send less or none at all due to the overflow.